Refactor - Reorganization of project
Renamed static dir to static in the survey app, because
the media dir should contain user uploaded content. Moved
the template in portable app because we need the app's
templates to be in the app so user do not have to redo
everything. Removed custom javascript and updated to Bootstrap.
It means a lot of change for the most part simplification of
handling a lot of js library and css class.