Feat - Add Sankey generation in diagram
We use Pysankey (https://github.com/anazalea/pySankey),
it add a lot of dependancies but its a lot easier than
doing it ourselves in latex (It was a lot of pain but,
it was possible : https://tex.stackexchange.com/questions/40159/)

Also add some tests for Sankey diagram generation but As the
diagram creating a big survey with user take 8s to run we
commented it. If someone work on this specifically they will
probably see it.